Make using phpcs more user-friendly on local dev
This adds a new grunt task: local

Runs the same except the report for phpcs is full rather than check style.

I was not able to figure out a better way of extending the phpcs task beyond re-creating it, since the options in the original task include the report and reportFile.