Changed getDOMValue() to accept array.
$param = array(
  'get_text' => T/F <- should we get the text for the node?
  'selector' => string <- will be passed to filter() so the format is what the CssSelector expects
  'attr' => string <- default is value
);

So the only thing that's required is selector (there's no error handling here). attr and get_text are optional.

Added an example setup in ContactElement->setValues(). In this example, most of the elements just use the value attribute, but the final element (the text for the selected option) uses text().