Update our colorbox version
The problem with colorbox not being "found" by libraries is a result of a problem with the version regex changing.  "The version of the colorbox library could not be detected", leading us to a very baffling message that the library did not exist :|

I see that this update was made for colorbox in version 2.8 -- I updated our make file.

*grunt will need to be run after merged in