Added Critical security patch from Mike + CJ
The caching added in oa_core didn't take into consideration the user's uid when caching the parents and subspaces shown in the toolbar.  So if you are within a parent space and refresh the page for an admin user, all subspaces are shown (the second dropdown, not the Home dropdown).  If you then switch to an incognito window and refresh the same page, anonymous users will see Private Subspaces in that second dropdown menu.  If they try to visit the subspace they will get access_denied.

oa_core-parent-cache.patch to oa_core 2.21 will fix this issue.