IASC-159: Added visibility rules for panes that do not apply to IASC theme (mini panel toolbar)
ooking at page.tpl.php, it's clear that we are not using most of the items from this mini panel in our theme (It looks like we are using the menu). So we are just adding to the page array components for no reason.

I added visibility rules to the items in the Footer region of /admin/structure/mini-panels/list/oa_toolbar_panel/edit/content to only show these items if the theme is _not_ IASC.

These feature overrides have been added to IASC Configuration.