IASC-155: Finally fixed issue with time zone field location on node edit
I decided not to lock the feature component because it seemed less than ideal due to the size of the umbrella for the overrides.

After taking another look, I finally figured out the source for the funky placement of the time zone field on the node/edit page for Events. It turns out that the reason was because of the reappearance of the field_oa_media (Attachments) field.

Added update hook to delete this field from the oa_event instance.

We don't need that field on Meetings. Feature also removes the view mode setup that uses the field_oa_media as well.


Former-commit-id: 62f3c4e5d1eb3cce0e5d8fb92d41f763ec26e0d0