IASC-149: added view template for contact pane
Didn't realize that organization was also supposed to be in the right column. Since users can select the fields they want to display, there's no guarantee of wrapping the right col fields in a div. So added a flag and some display logic for the template for the contact pane.


Former-commit-id: dffb493d2911a8faff82083a6ac5cef6297f4226