IASC-421: Removed duplicate description widget from panelized announcement display
The node template incorrectly had two description widgets in the main content well. The first description widget displayed a summary/trimmed version of the contents of the node description/body. This widget was followed by a full output of the field.

I removed the summary version of the description widget because it is unnecessary.

Affected pages that will need to be reverted:
|Title|Template|Path|
|-----|----------|-----|
|L3 IASC System-wide response activations & deactivations|Announcement template|/iasc-transformative-agenda/news-public/l3-iasc-system-wide-response-activations-deactivations|
|HFTT achievements at mid-2015|Announcement template|/humanitarian-financing-task-team/news-public/hftt-achievements-mid-2015|


Former-commit-id: e859ba1f6b0c8982093052ad9ee52e99f34ed85a